ci: fix clang/MSVC -Werror warnings exposed by warnings-as-errors
Clean up warnings that become fatal under the release-werror preset.
gcc passed already; clang and MSVC exposed additional issues, several of
which CI never reached because the build stopped at the first failure.

- runtime_demo.c: cast RAND_MAX to float (implicit int->float value change)
- pipeline_harness.hpp: static_cast uint64_t bytes_transferred to double (C4244)
- csp_pipeline_demo.cpp: drop unused l3_capacity_/l2_capacity_ fields
- kpu-calibrate.cpp: drop dead 'submitted' counter; [[maybe_unused]] ROW_BITS
- kpu-validate.cpp: [[maybe_unused]] ROW_BITS (documents address layout)
- benchmark.cpp: [[maybe_unused]] unused GREEN/CYAN/YELLOW color constants
- dma_harness.hpp: [[maybe_unused]] include_noc_ field
- verify_depthwise_conv.cpp: drop unused ElementwiseDescriptor e6
- kpu_native.cpp: avoid braced-scalar-init for 1-element array shape

Verified clean with both gcc and clang under release-werror locally.

+++ examples/c/runtime_demo.c
@@ -56,7 +56,7 @@ static const char* format_bytes(KPUSize bytes, char* buf, size_t buf_size) {
 /* Helper: fill array with random values */
 static void fill_random(float* data, size_t count) {
     for (size_t i = 0; i < count; i++) {
-        data[i] = ((float)rand() / RAND_MAX) * 2.0f - 1.0f;
+        data[i] = ((float)rand() / (float)RAND_MAX) * 2.0f - 1.0f;
     }
 }
